news_aggregator_app/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib[bcrypt]
requests
apscheduler
pydantic
email-validator
pytest
python-dotenv